+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Update the datasets used by GLAD alert emails to the correct ones. [#173192978](https://www.pivotaltracker.com/story/show/173192978)
 
+- Add cron schedule entry for GLAD alerts. [#173192978](https://www.pivotaltracker.com/story/show/173192978)
+
 # v.1.4.2
 
 ## 19/05/2020