+++ CHANGELOG.md
@@ -1,11 +1,15 @@
+# v.1.5.1
+
+## 03/07/2020
+
+- Add `redirect=false` via query param to disable redirection in re-send confirmation endpoint.
+
 # v.1.5.0
 
 ## 05/06/2020
 
 - Add endpoint for testing subscription email alerts for GLAD alerts and VIIRS Fires. [#172772548](https://www.pivotaltracker.com/story/show/172772548)
-
 - Update the datasets used by GLAD alert emails to the correct ones. [#173192978](https://www.pivotaltracker.com/story/show/173192978)
-
 - Add cron schedule entry for GLAD alerts. [#173192978](https://www.pivotaltracker.com/story/show/173192978)
 
 # v.1.4.2